[IMP] travis2docker: Add support for key to specify PostgreSQL version
This change adds support for the key `addons` -> `postgresql`, located
in the file `.travis.yml`. That key is used to manually specify the
PostgreSQL version, instead of using the default one.

In addition, this change adds a new test case, a long with a new example
file, to cover that new feature.

For more information, see:
https://docs.travis-ci.com/user/database-setup/#Using-a-different-PostgreSQL-Version

Closes #116