[IMP] travis2docker: Ensure changes in DB are saved before exiting
The current approach to ensure all database changes are saved before the
container is stopped, is to wait two seconds, so PostgreSQL has enough
time to write changes to disk. However, this approach does not guarantee
changes will be saved safely, especially when dealing with unlogged
tables, which will not be written until the PostgreSQL service is
stopped.

This commit changes the approach. Instead of waiting a prudent amount of
time, this causes the PostgreSQL to be stopped before exiting the
container.