[IMP] allow using sudo in the OCA runbot
Current situation; the install section of travis is mapped to
instructions running with a non root user.

However, some setups in OCA projects such as l10n-brazil, are using
sudo in their travis file to run things such as locales updates, which
fails (blocked on password prompt).

With this change, we parse the lines starting with sudo in the section
and place them in a different script, called sudo_runs. This script is executed
*after* the runs script, in a section where the user is root.

known limitation: this does not work if the commands requiring sudo and the others
must be run in a specific order. But it should be good enough for oca/l10n-brazil.