[FIX] travis2docker: Always exit with a failure code
The program's exit code was being taken from the list that contains the
created scripts. When `os.exit` gets a parameter different from `None` and
an integer, the parameter is printed and the exit status will be one (i.e.
failure).

This was structured that way to be able to print created scripts when called
from `__main__`.

To solve the above, resulting scripts are always printed and no arguments are
passed as exit code.

In addition, printer result is prettier now, it's as follows::

```
Generated scripts:
- script_1
- script_2
```

Closes #99