[FIX] t2d: Remove vim YouCompleteMe plugin I#165 (#184)
YouCompleteMe was causing issues as ycmd was not being installed properly,
however installating it offers no benefit since it did not mention
support for python3 autocompleting, and it is not used for the
pdb + TAB -> __import__('pdb').set_trace() feature which is the most
requested feature by Vauxoo Devs. Therefore it was removed.