[FIX] support START_SSH, START_ODOO and START_PSQL args
Fixes #200.

When supplying t2d the "START_SSH" argument during setup it is expected
that the generated image will automatically start running  the ssh
service.

This was not happening when using --deployv, it has been fixed by
adding the logic to autostart the server in the entrypoint for it.

Since the SSH server can only be started by root, the other commands
in the entrypoint have been modified to use the correct user if the
image is being run as root.

The same support has been given to ODOO and PSQL services. They start by
default but this can be avoided by providing a falsy value for
START_ODOO and START_PSQL respectively.