[CI] Update read the docs configuration
A configuration file fo RTD has been added, it uses the new key
`build.os` instead of the deprecated `build.image` one. The html theme
will also be now set regardless of whether it runs in RTD or locally, as
failing to do this generated errors on RTD builds (on RTD itself).

Closes #205.