[REV] add .travis.yml back
Previous CI and tox runs apparently lied. This is because tests run
against the main branch upstream, and not against the source code.

This generates false positives/negatives that are only found when a
merge completes but spending time fixing it is not a priority as of now.
Therefore the quickfix of adding the file back has been chosen.