[IMP] travis2docker: fix permissions and simplify shell in DeployV
The entrypoint used for DeployV-based projects now automatically fixes
ownership for files in `/home/odoo` that cannot be adjusted at build
time because some paths are declared as volumes in the parent DeployV
images (e.g. `~/.ssh`).

It also starts an interactive shell (`bash`) when `START_ODOO=0`, so
it is no longer required to override the entrypoint with
`--entrypoint bash` when running `t2d` for development purposes.

Closes #217