[FIX] travis2docker: don't overwrite existing keys (#225)
When granting access to SSH keys by adding the same public key to
the `authorized_keys` file, the existing file was being overwritten,
because the mode "w" was being used instead of "a". This is the
equivalent of using `>` instead of `>>` in shell.

To solve this, the "a" mode is now used.